auth-service/requirements.txt
Flask
PyJWT
requests
werkzeug
python-dotenv
